Added a bunch of loading feedback for the user

+++ src/lib/components/accountComponents/ArchiveAccountPopup.svelte
@@ -5,12 +5,16 @@
   import { AccountHierarchy } from '../../../params/accountHierarchy'
   import { goto, invalidate } from '$app/navigation'
   import Alert from '$lib/components/sharedComponents/Alert.svelte'
+  import LoadingSpinner from '$lib/components/sharedComponents/LoadingSpinner.svelte'
 
   export let accountId: number
   export let isParent: boolean
   let error: string
+  let isLoading = false
 
   async function onClick () {
+    if (isLoading) return
+    isLoading = true
     const endpoint = isParent ? AccountHierarchy.PARENT_ACCOUNT : AccountHierarchy.ACCOUNT
     const res = await fetch(`/${endpoint}/${accountId}`, {
       method: 'DELETE',
@@ -26,10 +30,12 @@
       const body = await res.json()
       error = body.error ?? body.message
     }
+    isLoading = false
   }
   function onExit () {
     error = ''
     $openPopup = false
+    isLoading = false
   }
 </script>
 
@@ -40,7 +46,7 @@
     <Alert>{error}</Alert>
   {/if}
   <div>
-    <Button on:click={onClick}>Yes</Button>
+    <Button disabled={isLoading} on:click={onClick}>Yes{#if isLoading}<LoadingSpinner/>{/if}</Button>
     <Button secondary={true} on:click={onExit}>No</Button>
   </div>
 </Popup>
